Move get_navigation_menu_items to a new wagtail.wagtailadmin.navigation module
